+++ constants.py
@@ -67,10 +67,10 @@ PRIMARY_MONITOR = 0
 
 # The model you are using, to calculate how many tokens the current message is
 # Ensure this is correct! Used for token count estimation
-MODEL = "nvidia/nemotron-3-nano"
+MODEL = "qwen/qwen3-vl-8b"
 
 # Context size (maximum number of tokens in the prompt) Will target upto 90% usage of this limit
-CONTEXT_SIZE = 75000
+CONTEXT_SIZE = 32000
 
 # This is your name
 #UNIQUE#